requirements.txt
fastapi
pydantic
pydantic-settings
uvicorn[standard]
python-multipart
azure-core
azure-ai-documentintelligence
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-storage
PyPDF2
pandas
openpyxl
python-dotenv
pillow
pymupdf
opencv-python
pytesseract
